+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased] Unreleased
 
+## [4.2.0] 2024-05-24;
+
 ### Added
 
 - Allow plugins to be loaded from arbitrary absolute or relative paths in the `WPLoader` module.
@@ -2189,4 +2191,5 @@ Read the [migration guide here](https://wpbrowser.wptestkit.dev/migration/from-v
 [4.1.7]: https://github.com/lucatume/wp-browser/compare/4.1.6...4.1.7
 [4.1.8]: https://github.com/lucatume/wp-browser/compare/4.1.7...4.1.8
 [4.1.9]: https://github.com/lucatume/wp-browser/compare/4.1.8...4.1.9
-[unreleased]: https://github.com/lucatume/wp-browser/compare/4.1.9...HEAD
+[4.2.0]: https://github.com/lucatume/wp-browser/compare/4.1.9...4.2.0
+[unreleased]: https://github.com/lucatume/wp-browser/compare/4.2.0...HEAD